+++ jgclark.Dashboard/src/index.js
@@ -8,7 +8,12 @@
 /**
  * Imports
  */
+import pluginJson from '../plugin.json'
 import { clo, JSP, logDebug, logError, logInfo, logWarn } from '@helpers/dev'
+// No longer used
+// import {  getNotesWithTagOrMention} from './tagMentionCache'
+
+const pluginID = 'jgclark.Dashboard'
 
 /**
  * Command Exports
@@ -46,7 +51,10 @@ export {
   reactWindowInitialised,
 } from './reactMain.js'
 
-export { onUpdateOrInstall, init, onSettingsUpdated, versionCheck } from './NPHooks'
+export {
+  // onUpdateOrInstall,
+  init, onSettingsUpdated, versionCheck
+} from './NPHooks'
 
 export {
   generateTagMentionCache,
@@ -56,16 +64,48 @@ export {
 export { externallyStartSearch } from './dataGenerationSearch.js'
 
 //-----------------------------------------------------------------------------
-// TODO(later): remove this for testing tag cache
 
-import {
-  getNotesWithTagOrMention
-} from './tagMentionCache'
+import { cleanDashboardSettings } from './perspectiveHelpers'
+import { renameKeys } from '@helpers/dataManipulation'
+import { saveSettings } from '@helpers/NPConfiguration'
+
+// Migrate some setting names to new names
+// TODO(later): remove this code in v2.3.0+
+export async function onUpdateOrInstall(): Promise<void> {
+  try {
+    logInfo(pluginJson, `onUpdateOrInstall() starting ...`)
+    const initialSettings = DataStore.settings
+    // clo(initialSettings, `onUpdateOrInstall - initialSettings:`)
+
+    // Migrate some setting names to new names
+    const keysToChange = {
+      usePerspectives: 'perspectivesEnabled',
+      includeFolderName: 'showFolderName',
+      includeScheduledDates: 'showScheduledDates',
+      includeTaskContext: 'showTaskContext',
+    }
+    logInfo(pluginJson, `onUpdateOrInstall() renaming any necessary keys from 2.1.x to 2.2.x ...`)
+    const migratedSettings = renameKeys(initialSettings, keysToChange)
+    clo(migratedSettings, `onUpdateOrInstall - migratedSettings:`)
+
+    // Save the settings back to the DataStore
+    if (migratedSettings !== initialSettings) {
+      const result = await saveSettings(pluginID, migratedSettings)
+      logInfo(`onUpdateOrInstall`, `Changes detected. Saved settings with result: ${JSP(result)}`)
+    }
+    logInfo(pluginJson, `onUpdateOrInstall() finished.`)
+  } catch (err) {
+    logError(pluginJson, `onUpdateOrInstall() error: ${err.message}`)
+  }
+}
+
+//-----------------------------------------------------------------------------
 
 /**
  * Test functions
  */
-export async function testTagCache(): Promise<void> {
-  let res = await getNotesWithTagOrMention(['@home'], false)
-  res = await getNotesWithTagOrMention(['#jgcDR'], false)
-}
+// No longer used
+// export async function testTagCache(): Promise<void> {
+//   let res = await getNotesWithTagOrMention(['@home'], false)
+//   res = await getNotesWithTagOrMention(['#jgcDR'], false)
+// }